"Can not Read/Write files from Python Template" For UWP apps, by default the current directory is the package directory, which is readonly.  The only writeable directory is the "Local app data" directory, Users\DefaultAccount\AppData\Local\Packages\<package name>\LocalState.
To help the user in this scenario, we are adding back the os.chdir() and os.getcwd() functions (which uses SetCurrentDirectoryW() and GetCurrentDirectoryW() respectively).  While fixing this bug, I noticed that that sys.prefix and sys.path were not set properly.  Thus, when the user changes the current directory, importing modules breaks.  All the python standard libraries now lives under a new directory called PythonHome under the application directory.  Now, PythonHome, sys.prefix, and sys.path points to this directory.

After this change, the user can add the following at the beginning of the script to change the current directory to point to a writeable directory:

import os;
os.chdir(os.environ["LOCALAPPDATA"])